+++ doc/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 Version 1.1
 -----------
 
+* Add a new "command" command to measure the timing of a program
 * Fix Windows 32bit issue on Python 2.7, fix by yattom.
 
 Version 1.0 (2017-03-17)